+++ app/src/main/java/com/example/anywherefitness/ui/WalkthroughActivity.kt
@@ -8,6 +8,7 @@ import androidx.lifecycle.ViewModelProviders
 import com.example.anywherefitness.R
 import com.example.anywherefitness.model.User
 import com.example.anywherefitness.ui.Instructor.InstructorActivity
+import com.example.anywherefitness.ui.client.ClientActivity
 import com.example.anywherefitness.viewmodel.LoginViewModel
 import kotlinx.android.synthetic.main.activity_walkthrough.*
