+++ app/src/main/java/com/example/anywherefitness/ui/client/FindClassesFragment.kt
@@ -19,6 +19,7 @@ import com.example.anywherefitness.R
 import com.example.anywherefitness.database.UserRepo
 import com.example.anywherefitness.model.FitnessClass
 import com.example.anywherefitness.ui.LoginActivity
+import com.example.anywherefitness.viewmodel.FindClassesViewModel
 import kotlinx.android.synthetic.main.fitness_class_list_view.view.*
 import kotlinx.android.synthetic.main.fragment_find_classes.*
